+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/command/AbstractCommand.java
@@ -2,7 +2,7 @@ package com.agonyforge.mud.demo.cli.command;
 
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
-import com.agonyforge.mud.demo.cli.CommandException;
+import com.agonyforge.mud.demo.cli.question.CommandException;
 import com.agonyforge.mud.demo.cli.RepositoryBundle;
 import com.agonyforge.mud.models.dynamodb.impl.MudCharacter;
 import com.agonyforge.mud.models.dynamodb.impl.MudItem;
