+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/command/ScoreCommand.java
@@ -4,7 +4,7 @@ import com.agonyforge.mud.core.cli.Question;
 import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
-import com.agonyforge.mud.demo.cli.CharacterSheetFormatter;
+import com.agonyforge.mud.demo.cli.question.CharacterSheetFormatter;
 import com.agonyforge.mud.demo.cli.RepositoryBundle;
 import com.agonyforge.mud.models.dynamodb.impl.MudCharacter;
 import com.agonyforge.mud.models.dynamodb.service.CommService;
