+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/question/CharacterEffortQuestion.java
@@ -1,4 +1,4 @@
-package com.agonyforge.mud.demo.cli;
+package com.agonyforge.mud.demo.cli.question;
 
 import com.agonyforge.mud.core.cli.Color;
 import com.agonyforge.mud.core.cli.Question;
@@ -6,11 +6,12 @@ import com.agonyforge.mud.core.cli.Response;
 import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
+import com.agonyforge.mud.demo.cli.RepositoryBundle;
 import com.agonyforge.mud.demo.cli.menu.MenuItem;
 import com.agonyforge.mud.demo.cli.menu.MenuPane;
 import com.agonyforge.mud.demo.cli.menu.MenuPrompt;
 import com.agonyforge.mud.demo.cli.menu.MenuTitle;
-import com.agonyforge.mud.models.dynamodb.constant.Stat;
+import com.agonyforge.mud.models.dynamodb.constant.Effort;
 import com.agonyforge.mud.models.dynamodb.impl.MudCharacter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,17 +22,17 @@ import java.util.Arrays;
 import java.util.Locale;
 
 @Component
-public class CharacterStatQuestion extends BaseQuestion {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CharacterStatQuestion.class);
-    private static final int STARTING_STATS = 6;
+public class CharacterEffortQuestion extends BaseQuestion {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CharacterEffortQuestion.class);
+    private static final int STARTING_EFFORTS = 4;
 
     private final MenuPane menuPane = new MenuPane();
 
-    public CharacterStatQuestion(ApplicationContext applicationContext,
-                                 RepositoryBundle repositoryBundle) {
+    public CharacterEffortQuestion(ApplicationContext applicationContext,
+                                   RepositoryBundle repositoryBundle) {
         super(applicationContext, repositoryBundle);
 
-        menuPane.setTitle(new MenuTitle("Allocate Stat Points"));
+        menuPane.setTitle(new MenuTitle("Allocate Effort Bonuses"));
         menuPane.setPrompt(new MenuPrompt());
     }
 
@@ -47,19 +48,19 @@ public class CharacterStatQuestion extends BaseQuestion {
 
     @Override
     public Response answer(WebSocketContext webSocketContext, Input input) {
-        String nextQuestion = "characterStatQuestion";
+        String nextQuestion = "characterEffortQuestion";
         Output output = new Output();
         MudCharacter ch = getCharacter(webSocketContext, output).orElseThrow();
         String choice = input.getInput().toUpperCase(Locale.ROOT);
-        int totalPoints = computeStatPoints(ch);
+        int totalPoints = computeEffortPoints(ch);
 
         if (choice.contains("+")) {
-            if (totalPoints >= STARTING_STATS) {
+            if (totalPoints >= STARTING_EFFORTS) {
                 output.append("[red]You don't have any more points to allocate!");
             } else {
                 try {
                     int i = Integer.parseInt(choice.substring(0, choice.length() - 1)) - 1;
-                    ch.addStat(Stat.values()[i], 1);
+                    ch.addEffort(Effort.values()[i], 1);
                 } catch (NumberFormatException | ArrayIndexOutOfBoundsException e) {
                     output.append("[red]Oops! Try a number with a plus or minus!");
                 }
@@ -70,18 +71,18 @@ public class CharacterStatQuestion extends BaseQuestion {
             } else {
                 try {
                     int i = Integer.parseInt(choice.substring(0, choice.length() - 1)) - 1;
-                    ch.addStat(Stat.values()[i], -1);
+                    ch.addEffort(Effort.values()[i], -1);
                 } catch (NumberFormatException | ArrayIndexOutOfBoundsException e) {
-                    output.append("[red][red]Oops! Try a number with a plus or minus!");
+                    output.append("[red]Oops! Try a number with a plus or minus!");
                 }
             }
         } else {
             if (choice.equals("S")) {
-                if (totalPoints == STARTING_STATS) {
-                    output.append("[green]Character stats saved!");
-                    nextQuestion = "characterEffortQuestion";
+                if (totalPoints == STARTING_EFFORTS) {
+                    output.append("[green]Character efforts saved!");
+                    nextQuestion = "characterMenuQuestion";
                 } else {
-                    output.append("[red]Please allocate exactly %d points for your stats.", STARTING_STATS);
+                    output.append("[red]Please allocate exactly %d points for your efforts.", STARTING_EFFORTS);
                 }
             } else {
                 output.append("[red]Oops! Try a number with a plus or minus!");
@@ -95,23 +96,23 @@ public class CharacterStatQuestion extends BaseQuestion {
         return new Response(next, output);
     }
 
-    private int computeStatPoints(MudCharacter ch) {
-        return Arrays.stream(Stat.values())
-            .map(ch::getStat)
+    private int computeEffortPoints(MudCharacter ch) {
+        return Arrays.stream(Effort.values())
+            .map(ch::getEffort)
             .reduce(0, Integer::sum);
     }
 
     private void populateMenuItems(MudCharacter ch) {
         menuPane.getItems().clear();
 
-        int points = STARTING_STATS - computeStatPoints(ch);
+        int points = STARTING_EFFORTS - computeEffortPoints(ch);
 
-        menuPane.getItems().add(new MenuItem(" ", "[default]Enter the menu number and a plus (+) or minus (-) to add or subtract from a stat."));
-        menuPane.getItems().add(new MenuItem(" ", "[default]For example, '3+' to raise CON or '6-' to lower CHA."));
+        menuPane.getItems().add(new MenuItem(" ", "[default]Enter the menu number and a plus (+) or minus (-) to add or subtract from an effort."));
+        menuPane.getItems().add(new MenuItem(" ", "[default]For example, '1+' to raise 'Basic' or '5-' to lower 'Ultimate'."));
         menuPane.getItems().add(new MenuItem(" ", String.format("[default]Please allocate [white]%d more points [default]for your stats.", points)));
 
-        Arrays.stream(Stat.values())
-                .forEachOrdered(stat -> menuPane.getItems().add(new MenuItem((stat.ordinal() + 1) + "[+/-]", String.format("%15s (%d)", stat.getName(), ch.getStat(stat)))));
+        Arrays.stream(Effort.values())
+            .forEachOrdered(effort -> menuPane.getItems().add(new MenuItem((effort.ordinal() + 1) + "[+/-]", String.format("%15s (%d)", effort.getName(), ch.getEffort(effort)))));
 
         menuPane.getItems().add(new MenuItem("S", "Save"));
     }
