+++ agonyforge-mud-demo/src/main/java/com/agonyforge/mud/demo/cli/question/CommandQuestion.java
@@ -1,10 +1,11 @@
-package com.agonyforge.mud.demo.cli;
+package com.agonyforge.mud.demo.cli.question;
 
 import com.agonyforge.mud.core.cli.Question;
 import com.agonyforge.mud.core.cli.Response;
 import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
+import com.agonyforge.mud.demo.cli.RepositoryBundle;
 import com.agonyforge.mud.demo.cli.command.Command;
 import com.agonyforge.mud.models.dynamodb.impl.CommandReference;
 import com.agonyforge.mud.models.dynamodb.impl.MudCharacter;
