+++ agonyforge-mud-demo/src/test/java/com/agonyforge/mud/demo/cli/question/CommandQuestionTest.java
@@ -1,11 +1,14 @@
-package com.agonyforge.mud.demo.cli;
+package com.agonyforge.mud.demo.cli.question;
 
 import com.agonyforge.mud.core.cli.Question;
 import com.agonyforge.mud.core.cli.Response;
 import com.agonyforge.mud.core.web.model.Input;
 import com.agonyforge.mud.core.web.model.Output;
 import com.agonyforge.mud.core.web.model.WebSocketContext;
+import com.agonyforge.mud.demo.cli.RepositoryBundle;
 import com.agonyforge.mud.demo.cli.command.Command;
+import com.agonyforge.mud.demo.cli.question.CommandException;
+import com.agonyforge.mud.demo.cli.question.CommandQuestion;
 import com.agonyforge.mud.models.dynamodb.impl.CommandReference;
 import com.agonyforge.mud.models.dynamodb.impl.MudCharacter;
 import com.agonyforge.mud.models.dynamodb.repository.CommandRepository;
